+++ docs/tutorial/examples/extut03_building_blocks.py
@@ -176,6 +176,10 @@ boundaries["left"] = fem.Boundary(displacement, fx=f0)
 boundaries["right"] = fem.Boundary(displacement, fx=f1, skip=(1, 0, 0))
 boundaries["move"] = fem.Boundary(displacement, fx=f1, skip=(0, 1, 1), value=0.5)
 
+plotter = boundaries["left"].plot(color="green")
+plotter = boundaries["right"].plot(plotter=plotter, color="red")
+boundaries["move"].plot(plotter=plotter, color="blue", point_size=5).show()
+
 # %%
 # Partition of deegrees of freedom
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~